Event member

+++ src/main/java/com/eventplanner/service/EventService.java
@@ -6,11 +6,13 @@
 package com.eventplanner.service;
 
 import com.eventplanner.domain.Event;
+import com.eventplanner.domain.EventUser;
 import com.eventplanner.repo.EventRepository;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 
 /**
@@ -25,17 +27,26 @@ public class EventService {
     @Autowired
     private EventGroupService groupService;
     @Autowired
-    private EventMemberService memberService; 
+    private EventUserService userService;
 
+    /*
+    LocalDateTime is converted in the controller from the HTML localdate object
+    GOTO Event class for more info
+    Checks for duplicates
+    */
     public Event createEvent(String name, Integer groupId, String description, String location,
             LocalDateTime startDate, LocalDateTime endDate) {
-        //Add duplicate name here
-        return eventRepo.save(new Event(name, groupService.findGroupById(groupId), description, location,
-                startDate, endDate));
+        if (eventRepo.findByNameAndEventGroup(name, groupService.findGroupById(groupId)) == null) {
+            return eventRepo.save(new Event(name, groupService.findGroupById(groupId), description, location,
+                    startDate, endDate));
+        }
+        return null;
     }
 
     public void deleteEvent(Integer eventId) {
-        eventRepo.delete(eventId);
+        if(eventRepo.exists(eventId)){
+            eventRepo.delete(eventId);
+        }
     }
 
     public void updateEvent(Integer eventId, String name, String description, String location,
@@ -49,18 +60,6 @@ public class EventService {
         event.setEndDate(endYear, endMonth, endDay, endHour, endMinute);
         eventRepo.save(event);
     }
-    
-    public List findMemberEvents(List<Integer> eventList){
-        List<Event> list = new ArrayList();
-        eventList.forEach(item -> list.add(eventRepo.findOne(item)));
-        return list;
-    }
-    
-    public List findGroupEvents(Integer groupId, List<Integer> eventList){
-        List<Event> list = new ArrayList();
-        eventList.forEach(item -> list.add(eventRepo.findByIdAndEventGroup(item,groupService.findGroupById(groupId))));
-        return list;
-    }
 
     public List<Event> getAllEvents() {
         List<Event> list = new ArrayList();
@@ -68,15 +67,23 @@ public class EventService {
         return list;
     }
 
+    /*
+    Adds and removes members uniquely to event and not group
+    Checks for duplicates
+    */
     public Event addMember(Integer eventId, String username) {
         Event event = eventRepo.findOne(eventId);
-        event.getEventMemberList().add(memberService.findByUsername(username));
+        EventUser user = userService.findByUsername(username);
+        if (user.getAuthorities().contains(new SimpleGrantedAuthority("USER"))
+                && !event.getEventMemberList().contains(user)) {
+            event.getEventMemberList().add(user);
+        }
         return eventRepo.save(event);
     }
-    
-    public Event removeMember(Integer eventId, String username){
+
+    public Event removeMember(Integer eventId, String username) {
         Event event = eventRepo.findOne(eventId);
-        event.getEventMemberList().remove(memberService.findByUsername(username));
+        event.getEventMemberList().remove(userService.findByUsername(username));
         return eventRepo.save(event);
     }
 }
